+++ v20.2/build-a-go-app-with-cockroachdb-gorm.md
@@ -21,27 +21,13 @@ For another use of GORM with CockroachDB, see our [`examples-orms`](https://gith
 
 {% include {{page.version.version}}/app/before-you-begin.md %}
 
-## Step 1. Install the GORM ORM
-
-To install [GORM](http://gorm.io), run the following commands:
-
-{% include copy-clipboard.html %}
-~~~ shell
-$ go get -u github.com/lib/pq # dependency
-~~~
-
-{% include copy-clipboard.html %}
-~~~ shell
-$ go get -u github.com/jinzhu/gorm
-~~~
-
 <section class="filter-content" markdown="1" data-scope="secure">
 
-## Step 2. Create the `maxroach` user and `bank` database
+## Step 1. Create the `maxroach` user and `bank` database
 
 {% include {{page.version.version}}/app/create-maxroach-user-and-bank-database.md %}
 
-## Step 3. Generate a certificate for the `maxroach` user
+## Step 2. Generate a certificate for the `maxroach` user
 
 Create a certificate and key for the `maxroach` user by running the following command. The code samples will run as this user.
 
@@ -50,17 +36,17 @@ Create a certificate and key for the `maxroach` user by running the following co
 $ cockroach cert create-client maxroach --certs-dir=certs --ca-key=my-safe-directory/ca.key
 ~~~
 
-## Step 4. Run the Go code
+## Step 3. Run the Go code
 
-The following code uses the [GORM](http://gorm.io) ORM to map Go-specific objects to SQL operations. Specifically:
+The following code uses the [GORM](http://gorm.io) ORM (v1) to map Go-specific objects to SQL operations, and the [`crdbgorm`](https://godoc.org/github.com/cockroachdb/cockroach-go/crdb/crdbgorm) package to handle [transactions](transactions.html). Specifically:
 
 - `db.AutoMigrate(&Account{})` creates an `accounts` table based on the Account model.
 - `db.Create(&Account{})` inserts rows into the table.
 - `db.Find(&accounts)` selects from the table so that balances can be printed.
-- The funds transfer occurs in `transferFunds()`. To ensure that we [handle retry errors](transactions.html#client-side-intervention), we write an application-level retry loop that, in case of error, sleeps before trying the funds transfer again. If it encounters another error, it sleeps again for a longer interval, implementing [exponential backoff](https://en.wikipedia.org/wiki/Exponential_backoff).
+- The funds transfer occurs in `transferFunds()`. To ensure that we [handle retry errors](transactions.html#client-side-intervention), we wrap the function call in [`crdbgorm.ExecuteTx()`](https://github.com/cockroachdb/cockroach-go/blob/master/crdb/crdbgorm/gorm.go#L29).
 
-Copy the code or
-<a href="https://raw.githubusercontent.com/cockroachdb/docs/master/_includes/{{ page.version.version }}/app/gorm-sample.go" download>download it directly</a>.
+Copy the code below and save it to a file named `main.go`, or
+<a href="https://raw.githubusercontent.com/cockroachdb/docs/master/_includes/{{ page.version.version }}/app/gorm-basic-sample/main.go" download>download it directly</a>.
 
 {{site.data.alerts.callout_success}}
 To clone a version of the code below that connects to insecure clusters, run the command below. Note that you will need to edit the connection string to use the certificates that you generated when you set up your secure cluster.
@@ -70,14 +56,21 @@ To clone a version of the code below that connects to insecure clusters, run the
 
 {% include copy-clipboard.html %}
 ~~~ go
-{% include {{ page.version.version }}/app/gorm-sample.go %}
+{% include {{ page.version.version }}/app/gorm-basic-sample/main.go %}
+~~~
+
+Use `go mod init` to get the dependencies:
+
+{% include copy-clipboard.html %}
+~~~ shell
+$ go mod init hello-world-go-gorm
 ~~~
 
 Then run the code:
 
 {% include copy-clipboard.html %}
 ~~~ shell
-$ go run gorm-sample.go
+$ go run main.go
 ~~~
 
 The output should show the account balances before and after the funds transfer:
@@ -95,18 +88,18 @@ Balance at '2019-08-06 13:37:19.325654 -0400 EDT m=+0.048303286':
 
 <section class="filter-content" markdown="1" data-scope="insecure">
 
-## Step 2. Create the `maxroach` user and `bank` database
+## Step 1. Create the `maxroach` user and `bank` database
 
 {% include {{page.version.version}}/app/insecure/create-maxroach-user-and-bank-database.md %}
 
-## Step 3. Run the Go code
+## Step 2. Run the Go code
 
-The following code uses the [GORM](http://gorm.io) ORM to map Go-specific objects to SQL operations. Specifically:
+The following code uses the [GORM](http://v1.gorm.io/) ORM (v1) to map Go-specific objects to SQL operations, and the [`crdbgorm`](https://godoc.org/github.com/cockroachdb/cockroach-go/crdb/crdbgorm) package to handle [transactions](transactions.html). Specifically:
 
 - `db.AutoMigrate(&Account{})` creates an `accounts` table based on the Account model.
 - `db.Create(&Account{})` inserts rows into the table.
 - `db.Find(&accounts)` selects from the table so that balances can be printed.
-- The funds transfer occurs in `transferFunds()`. To ensure that we [handle retry errors](transactions.html#client-side-intervention), we write an application-level retry loop that, in case of error, sleeps before trying the funds transfer again. If it encounters another error, it sleeps again for a longer interval, implementing [exponential backoff](https://en.wikipedia.org/wiki/Exponential_backoff).
+- The funds transfer occurs in `transferFunds()`. To ensure that we [handle retry errors](transactions.html#client-side-intervention), we wrap the function call in [`crdbgorm.ExecuteTx()`](https://github.com/cockroachdb/cockroach-go/blob/master/crdb/crdbgorm/gorm.go#L29).
 
 To get the code below, clone the `hello-world-go-gorm` repo to your machine:
 
@@ -117,10 +110,17 @@ git clone https://github.com/cockroachlabs/hello-world-go-gorm/
 
 {% include copy-clipboard.html %}
 ~~~ go
-{% include {{ page.version.version }}/app/insecure/gorm-sample.go %}
+{% include {{ page.version.version }}/app/insecure/gorm-basic-sample/main.go %}
 ~~~
 
-Change to the directory where you cloned the repo and run the code:
+Change to the directory where you cloned the repo, and use `go mod init` to get the dependencies:
+
+{% include copy-clipboard.html %}
+~~~ shell
+$ go mod init hello-world-go-gorm
+~~~
+
+Then run the code:
 
 {% include copy-clipboard.html %}
 ~~~ shell
